cart $ favorites done

+++ lib/presentations/common_blocs/cart/cart_state.dart
@@ -13,7 +13,7 @@ class CartLoading extends CartState {}
 /// Cart was loaded
 class CartLoaded extends CartState {
   final List<CartItemModel> cart;
-  final int priceOfGoods;
+  final double priceOfGoods;
 
   const CartLoaded({
     required this.cart,
